# epicash_api/src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple example NodeJS API project to demonstrate Swagger Documentation",
        "version": "1.0.0",
        "title": "Cash API",
        "contact": {
            "email": "abc@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:4212",
    "basePath": "/api",
    "paths" : {
...
        "/api/users/identify" : {
            "post" : {
                "summary" : "Identify user",
                "description": "Identify user",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "user object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email":{
                                    "type": "string"
                                },
                                "password":{
                                    "type": "string"
                                }
                            }

                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status":{
                                    "type": "boolean"
                                },
                                "id":{
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/api/users/addcash/{id}" : {
            "post" : {
                "summary" : "add money to user solde",
                "description": "add money to user solde",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "user object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/paye"

                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/users"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/api/users/paye/{id}" : {
            "post" : {
                "summary" : "Paye a list of product",
                "description": "Paye a list of product",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "user object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "card": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id":{
                                                "type":"integer"
                                            },
                                            "qty":{
                                                "type":"integer"
                                            }
                                        }
                                    }

                                    
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/returnPaye"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
    }, 
    "definitions": {
        "users": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "email":{
                    "type": "string"
                },
                "firstname":{
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "solde": {
                    "type": "integer"
                },
                "admin": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "date"
                },
                "createdAt": {
                    "type": "date"
                }
            }
        },
...
        "paye": {
            "type": "object",
            "properties": {
                "cash":{
                    "type": "integer"
                },
                "account":{
                    "type": "integer"
                },
                "cvv": {
                    "type": "integer"
                }
            }
        },
        "bills": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "reference": {
                    "type": "string"
                },
                "line": {
                    "type": "array",
                    "items":
                    {
                        "type": "object",
                        "properties": {
                            "fk_product": {
                                "type": "integer"
                            },
                            "qty": {
                                "type": "integer"
                            },
                            "productName": {
                                "type": "string"
                            },
                            "prix": {
                                "type": "integer"
                            },
                            "totalLine": {
                                "type": "integer"
                            }
                        }
                    }
                },
                "fk_user":{
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "createdAt":{
                    "type": "date"
                }
            }
        },
        "returnPaye":{
            "type": "object",
            "properties": {
                "status":{
                    "type": "boolean"
                },
                "data":{
                    "type":"object",
                    "properties": {
                        "missingSolde":{
                            "type": "integer"
                        },
                        "bills":{
                            "type":"object",
                            "$ref":"#/definitions/bills"
                        },
                        "user":{
                            "type":"object",
                            "$ref":"#/definitions/users"
                        }
                    }
                }
            }
        },
        "InvalidResponse": {
            "message":"Error message"
        }
    }
}